+++ src/MainLib/Marqdouj.DotNet.Web.JsInterop.Azure.Maps/Models/DataDrivenStyles/CoalesceGetStyleBuilder.cs
@@ -1,0 +1,53 @@
+﻿namespace Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Models.DataDrivenStyles
+{
+    internal class CoalesceGetStyleBuilder
+    {
+        private readonly List<string> _cases;
+
+        /// <summary>
+        /// Initializes a new instance of the CoalesceGetStyleBuilder class with the specified list of cases.
+        /// </summary>
+        /// <param name="cases"></param>
+        /// <param name="defaultCase"></param>
+        public CoalesceGetStyleBuilder(List<string> cases, string defaultCase = "")
+        {
+            ArgumentNullException.ThrowIfNull(cases, nameof(cases));
+            _cases = cases;
+            DefaultCase = defaultCase;
+        }
+
+        /// <summary>
+        /// Default case for the coalesce expression.
+        /// </summary>
+        public string DefaultCase { get; }
+
+        /// <summary>
+        /// Builds the coalesce style expression based on the defined cases.
+        /// </summary>
+        /// <returns></returns>
+        public object Build()
+        {
+            var result = new List<object> {
+                "coalesce",
+            };
+
+            foreach (var caseValue in _cases)
+            {
+                result.Add(new List<string> { "get", caseValue });
+            }
+
+            result.Add(DefaultCase);
+
+            return result;
+        }
+
+        /// <summary>
+        /// Returns a JSON string representation of the coalesce style expression built by this builder.
+        /// </summary>
+        /// <returns></returns>
+        public override string ToString()
+        {
+            return System.Text.Json.JsonSerializer.Serialize(Build());
+        }
+    }
+}